Pixel classification refactoring
Extracting measurements from overlay, and moving towards more functionality accessible from static methods.
