+++ qupath-experimental/src/main/java/qupath/lib/classifiers/gui/PixelClassifierGUI.java
@@ -50,7 +50,6 @@ import qupath.lib.roi.interfaces.PathShape;
 import qupath.lib.roi.interfaces.ROI;
 import qupath.opencv.processing.TypeAdaptersCV;
 
-import java.awt.Shape;
 import java.awt.geom.Area;
 import java.awt.geom.Path2D;
 import java.awt.image.BufferedImage;
